requirements.txt
typing==3.10.0.0
python-dotenv==0.15.0
xml.etree.ElementTree==4.5.0
lxml==4.6.3
zipfile36==0.1.3
requests==2.25.1
boto3==1.17.31
ntplib==0.3.4